Jugador-Equipo

+++ src/main/java/com/proyecto/wordle/dto/JugadorDTO.java
@@ -1,17 +1,17 @@
 package com.proyecto.wordle.dto;
 
+import com.proyecto.wordle.model.Equipo;
 import lombok.Setter;
 import lombok.Getter;
 
 @Setter
 @Getter
 public class JugadorDTO {
-    private Long idjugador;
+    private Integer idjugador;
     private Boolean admin;
     private String nombre;
     private String avatar;
     private Integer puntos;
-
-    private Long equipo_idequipo;
+    private Integer equipo_idequipo;
 
 }
