+++ src/main/java/com/proyecto/wordle/dto/JugadorModifDTO.java
@@ -7,10 +7,9 @@ import lombok.Setter;
 @Setter
 
 public class JugadorModifDTO {
-
     private Boolean admin;
     private String nombre;
     private String avatar;
     private Integer puntos;
-    private Long equipo_idequipo;
+    private Integer equipo_idequipo;
 }
